requirements.txt
pymemcache
python-memcached
pillow
boto3
